requirements.txt
telethon
pillow
lottie
opencv-python
cairosvg